+++ app/src/main/java/app/tryst/data/db/TrystDatabase.kt
@@ -56,7 +56,7 @@ import app.tryst.data.db.entity.ToyEntity
         EncounterPositionCrossRef::class,
         EncounterTagCrossRef::class,
     ],
-    version = 15,
+    version = 16,
     exportSchema = true,
 )
 @TypeConverters(Converters::class)